Move eval+emote picker to class
